#45 - Create `GeolocationButton` component
