+++ src/pages/search/LocationSelection.tsx
@@ -1,24 +1,12 @@
-import { useAppDispatch } from '@store';
-import { getWeatherByGeolocation } from '@slices/weatherSlice.thunks';
 import { LocationResults } from '@pages/search/LocationResults';
-import { ButtonWithIcon } from '@components/ButtonWithIcon';
+import { GeolocationButton } from '@pages/search/GeolocationButton';
 
-export const LocationSelection = () => {
-  const dispatch = useAppDispatch();
+export const LocationSelection = () => (
+  <div className="mt-2 grow space-y-4 px-3">
+    <GeolocationButton />
 
-  const getLocation = () => dispatch(getWeatherByGeolocation());
-
-  return (
-    <div className="mt-2 grow space-y-4 px-3">
-      <div className="mt-3 md:mt-5">
-        <ButtonWithIcon onClick={getLocation} icon="tabler:current-location">
-          Use your current location
-        </ButtonWithIcon>
-      </div>
-
-      <ul className="flex flex-col items-start gap-y-3">
-        <LocationResults />
-      </ul>
-    </div>
-  );
-};
+    <ul className="flex flex-col items-start gap-y-3">
+      <LocationResults />
+    </ul>
+  </div>
+);
